Explicit cast to allow building with GCC 2.95.4 20020320 on FreeBSD/Alpha 4.9.
git-svn-id: https://gxemul.svn.sourceforge.net/svnroot/gxemul/gxemul/trunk@61 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/main/StateVariable.cc
@@ -323,7 +323,7 @@ uint64_t StateVariable::ToInteger() const
 	case Bool:
 		return (*m_value.pbool)? 1 : 0;
 	case Double:
-		return *m_value.pdouble;
+		return (uint64_t) *m_value.pdouble;
 	case UInt8:
 		return *m_value.puint8;
 	case UInt16:
